feat(Feature): Turn advance search bar into a component
